literature_data_analysis/requirements.txt
langid == 1.1.5
nltk == 3.0.2
py2neo == 3.1.2
pymongo == 3.1
pyquery == 1.2.17
requests == 2.18.1
tqdm == 4.14.0
unicodecsv == 0.14.1
Unidecode == 0.04.17
PyYAML == 3.12
pika==0.10.0
